requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-graphql-jwt==0.3.4
django-model-utils==4.3.1
django-notifications-hq==1.7.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
frozenlist==1.3.3
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iso8601==1.1.0
jsonfield==3.1.0
msgpack==1.0.4
multidict==6.0.3
outcome==1.2.0
packaging==21.3
promise==2.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytz==2022.6
redis==4.3.4
Rx==1.6.1
selenium==4.6.0
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
swapper==1.3.0
text-unidecode==1.3
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
zope.interface==5.5.2
